+++ grafana-plugin/src/models/timezone/timezone.types.ts
@@ -1,0 +1,3 @@
+import { allTimezones } from './timezone.helpers';
+
+export type Timezone = (typeof allTimezones)[number];
